# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bounty Challenge API",
    "description": "API for the GitHub Issue Bounty Challenge - Rewards miners for valid bug reports",
    "version": "0.1.0",
    "contact": {
      "name": "PlatformNetwork",
      "url": "https://github.com/PlatformNetwork/bounty-challenge"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://chain.platform.network/api/v1/bridge/bounty-challenge",
      "description": "Production bridge (via Platform Network)"
    }
  ],
  "paths": {
...
    "/get_weights": {
      "get": {
        "summary": "Get miner weights",
        "description": "Returns current weights for all registered miners (platform-compatible)",
        "tags": ["Weights"],
        "parameters": [
          {
            "name": "epoch",
            "in": "query",
            "description": "Optional epoch number (defaults to current)",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Miner weights",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GetWeightsResponse" },
                "example": {
                  "weights": [
                    { "hotkey": "5FHneW46...", "weight": 0.25 },
                    { "hotkey": "5GrwvaEF...", "weight": 0.15 }
                  ],
                  "epoch": 12345678,
                  "challenge_id": "bounty-challenge",
                  "total_miners": 2
                }
              }
            }
          }
        }
      }
    },
...
    "/register": {
      "post": {
        "summary": "Register GitHub account",
        "description": "Link a GitHub username to a miner hotkey with cryptographic signature",
        "tags": ["Registration"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" },
              "example": {
                "hotkey": "5FHneW46xGXgs5mUiveU4sbTyGBzmstUspZC92UhjJM694ty",
                "github_username": "echobt",
                "signature": "0x1234abcd...",
                "timestamp": 1705622400
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registration result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RegisterResponse" },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "message": "Successfully registered @echobt with your hotkey.",
                      "error": null
                    }
                  },
                  "error": {
                    "value": {
                      "success": false,
                      "message": null,
                      "error": "Invalid signature. Make sure you're using the correct key."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/status/{hotkey}": {
      "get": {
        "summary": "Get miner status",
        "description": "Returns registration status and stats for a hotkey",
        "tags": ["Registration"],
        "parameters": [
          {
            "name": "hotkey",
            "in": "path",
            "required": true,
            "description": "Miner hotkey (SS58 format)",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Miner status",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StatusResponse" },
                "examples": {
                  "registered": {
                    "value": {
                      "registered": true,
                      "github_username": "echobt",
                      "valid_issues_count": 5,
                      "invalid_issues_count": 1,
                      "balance": 4,
                      "is_penalized": false,
                      "weight": 0.05
                    }
                  },
                  "not_registered": {
                    "value": {
                      "registered": false,
                      "github_username": null,
                      "valid_issues_count": null,
                      "invalid_issues_count": null,
                      "balance": null,
                      "is_penalized": false,
                      "weight": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/invalid": {
      "post": {
        "summary": "Record invalid issue",
        "description": "Record an issue as invalid (for penalty tracking)",
        "tags": ["Issues"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InvalidIssueRequest" },
              "example": {
                "issue_id": 123,
                "repo_owner": "PlatformNetwork",
                "repo_name": "bounty-challenge",
                "github_username": "spammer",
                "issue_url": "https://github.com/PlatformNetwork/bounty-challenge/issues/123",
                "issue_title": "Spam issue",
                "reason": "Low quality / spam"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InvalidIssueResponse" }
              }
            }
          }
        }
      }
    },
    "/issues": {
      "get": {
        "summary": "List all issues",
        "description": "Get cached GitHub issues with optional filters",
        "tags": ["Issues"],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "description": "Filter by state (open/closed)",
            "schema": { "type": "string", "enum": ["open", "closed"] }
          },
          {
            "name": "label",
            "in": "query",
            "description": "Filter by label (e.g., valid, invalid, bug)",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max results (default 50, max 200)",
            "schema": { "type": "integer", "default": 50 }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset",
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "List of issues",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/IssuesListResponse" },
                "example": {
                  "issues": [
                    {
                      "issue_id": 1,
                      "repo_owner": "PlatformNetwork",
                      "repo_name": "bounty-challenge",
                      "github_username": "echobt",
                      "title": "[BUG] Test issue",
                      "state": "open",
                      "labels": ["bug", "valid"],
                      "created_at": "2026-01-19T00:47:00Z",
                      "updated_at": "2026-01-19T00:47:00Z",
                      "closed_at": null,
                      "issue_url": "https://github.com/PlatformNetwork/bounty-challenge/issues/1"
                    }
                  ],
                  "count": 1,
                  "limit": 50,
                  "offset": 0
                }
              }
            }
          }
        }
      }
    },
    "/issues/pending": {
      "get": {
        "summary": "List pending issues",
        "description": "Get closed issues without valid/invalid label (need review)",
        "tags": ["Issues"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 50 }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "List of pending issues",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/IssuesListResponse" }
              }
            }
          }
        }
      }
    },
...
    "/hotkey/{hotkey}": {
      "get": {
        "summary": "Get hotkey details",
        "description": "Get detailed info for a miner hotkey including recent issues",
        "tags": ["Details"],
        "parameters": [
          {
            "name": "hotkey",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotkey details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HotkeyDetailsResponse" },
                "example": {
                  "hotkey": "5FHneW46...",
                  "github_username": "echobt",
                  "registered_at": "2026-01-15T10:00:00Z",
                  "valid_issues": 5,
                  "invalid_issues": 1,
                  "balance": 4,
                  "is_penalized": false,
                  "weight": 0.05,
                  "recent_issues": [
                    {
                      "issue_id": 1,
                      "repo_owner": "PlatformNetwork",
                      "repo_name": "bounty-challenge",
                      "title": "[BUG] Test",
                      "state": "open",
                      "labels": ["bug", "valid"],
                      "updated_at": "2026-01-19T00:47:00Z",
                      "issue_url": "https://github.com/..."
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Hotkey not found",
            "content": {
              "application/json": {
                "example": { "error": "Hotkey not found" }
              }
            }
          }
        }
      }
    },
...
    "/evaluate": {
      "post": {
        "summary": "Evaluate submission",
        "description": "Platform SDK evaluation endpoint (internal)",
        "tags": ["Platform SDK"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EvaluationRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evaluation result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EvaluationResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetWeightsResponse": {
        "type": "object",
        "properties": {
          "weights": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MinerWeight" }
          },
          "epoch": { "type": "integer" },
          "challenge_id": { "type": "string" },
          "total_miners": { "type": "integer" }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "required": ["hotkey", "github_username", "signature", "timestamp"],
        "properties": {
          "hotkey": { "type": "string", "description": "SS58 encoded hotkey" },
          "github_username": { "type": "string" },
          "signature": { "type": "string", "description": "Hex-encoded sr25519 signature" },
          "timestamp": { "type": "integer", "description": "Unix timestamp (must be within 5 min)" }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string", "nullable": true },
          "error": { "type": "string", "nullable": true }
        }
      },
...
      "EvaluationRequest": {
        "type": "object",
        "description": "Platform SDK evaluation request"
      },
...
    }
  },
...
}